handler/requirements.txt
openai-whisper
runpod
torch
torchaudio